Project reorganization - move WeatherDataAlert(s) to own file under Alerts in project.
