+++ WeatherApi/WeatherApi/DataSources/WeatherData.swift
@@ -168,26 +168,4 @@ struct ApiErrorData {
     var code: Int
     var message: String
 }
-
-// MARK: API Alerts
-struct WeatherDataAlerts {
-    let alerts: [WeatherDataAlert]
-}
-
-struct WeatherDataAlert: Identifiable {
-    let id: UUID
-    let category: String?
-    let msgtype: String?
-    let note: String?
-    let headline: String?
-    let effective: String?
-    let event: String?
-    let expires: String?
-    let desc: String?
-    let instruction: String?
-    let urgency: String?
-    let severity: String?
-    let areas: String?
-    let certainty: String?
-}
 // swiftlint:enable identifier_name
